Move csl_item_set_standard_citation to global workflow
