+++ manubot/pandoc_filter/tests/test_cite.py
@@ -94,36 +94,3 @@ def test_load_bibliography_from_text(path):
     assert csl_json[0]['title'] == 'Sci-Hub provides access to nearly all scholarly literature'
     if input_format in {'json', 'bib'}:
         assert csl_json[0].get('id') == 'doi:10.7554/elife.32822'
-
-
-@pytest.mark.parametrize(
-    ['csl_item', 'standard_citation'],
-    [
-        (
-            {'id': 'my-id', 'standard_citation': 'doi:10.7554/elife.32822'},
-            'doi:10.7554/elife.32822',
-        ),
-        (
-            {'id': 'doi:10.7554/elife.32822'},
-            'doi:10.7554/elife.32822',
-        ),
-        (
-            {'id': 'doi:10.7554/ELIFE.32822'},
-            'doi:10.7554/elife.32822',
-        ),
-        (
-            {'id': 'my-id'},
-            'raw:my-id',
-        ),
-    ],
-    ids=[
-        'from_standard_citation',
-        'from_doi_id',
-        'from_doi_id_standardize',
-        'from_row_id',
-    ]
-)
-def test_csl_item_set_standard_citation(csl_item, standard_citation):
-    output = csl_item_set_standard_citation(csl_item)
-    assert output is csl_item
-    assert output['standard_citation'] == standard_citation
